requirements.txt
langchain
python-dotenv
langchain-openai
beautifulsoup4
langchain_community
bs4
faiss-cpu
sentence-transformers
langchain_core
streamlit
